Split external scene change responsibilities by file
Give the tracker, snapshot session store, focus-return coordinator, and compile resolver explicit file ownership while preserving every class body, visibility, event hook, and mutable state owner.
