+++ Packages/src/Editor/FirstPartyTools/Compile/ExternalAssetFocusReturnService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e756216d6984843a9c45be220a9c78f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
