+++ Packages/src/Editor/FirstPartyTools/Compile/ExternalAssetSnapshotSessionStore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9c8a83b2ad24cb7800665ae367e6e14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
