+++ Packages/src/Editor/FirstPartyTools/Compile/ExternalSceneChangeTracker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6d5c1b8149e64d61b4dc57d174a8d1c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
